fixbug tail ddv
修复ddv日志监听
