+++ lib/cli/index.js
@@ -129,7 +129,7 @@ log.i18nInit().then(() => {
     argv.length = argv.length - 1
     argv.push('--help')
   }
-  commander.argv_source = argv
+  commander.argvSource = argv
   if (!argv || argv.length <= 2) {
     argv = argv.slice(0, 2)
     // 解析命令行输入
